some more reorganization
