+++ common-queries/src/main/scala/queries/schemas/odb/ObsQueries.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.common
+package queries.schemas.odb
 
 import cats.effect.Async
 import cats.implicits.*
@@ -32,7 +32,7 @@ import monocle.Focus
 import monocle.Getter
 import monocle.Lens
 import queries.common.ObsQueriesGQL.*
-import queries.schemas.implicits.*
+import queries.schemas.odb.conversions.*
 
 import java.time.Instant
 import scala.collection.immutable.SortedMap
