+++ common/src/main/scala/explore/common/UserPreferencesQueries.scala
@@ -35,8 +35,8 @@ import queries.schemas.UserPreferencesDB.Enums.*
 import queries.schemas.UserPreferencesDB.Scalars.*
 import queries.schemas.UserPreferencesDB.Types.LucumaObservationInsertInput
 import queries.schemas.UserPreferencesDB.Types.*
-import queries.schemas.WidthUpsertInput
-import queries.schemas.implicits.*
+import queries.schemas.odb.WidthUpsertInput
+import queries.schemas.odb.conversions.*
 import react.gridlayout.{BreakpointName => _, _}
 import reactST.highcharts.highchartsStrings.chart_
