+++ explore/src/main/scala/explore/observationtree/AsterismGroupObsListActions.scala
@@ -22,7 +22,7 @@ import lucuma.core.model.Target
 import lucuma.schemas.ObservationDB
 import lucuma.schemas.ObservationDB.Types.*
 import queries.common.TargetQueriesGQL
-import queries.schemas.implicits.*
+import queries.schemas.odb.conversions.*
 
 import scala.annotation.unused
 import scala.collection.immutable.SortedSet
