+++ workers/src/main/scala/workers/itc/ITCGraphRequests.scala
@@ -25,7 +25,7 @@ import lucuma.refined.*
 import org.typelevel.log4cats.Logger
 import queries.common.ITCQueriesGQL.*
 import queries.schemas.ITC
-import queries.schemas.itc.implicits.*
+import queries.schemas.itc.conversions.*
 import workers.*
 
 import java.util.UUID
